+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+2.0 (2016-xx-xx)
+----------------
+
+- **BACKWARD-INCOMPATIBLE** Removed the ``HostsMiddleware``, deprecated in
+  django-hosts 1.0.
+
 1.4 (2016-01-21)
 ----------------